refactor(ta): extract shared TaTimeEntryRow and TaParticipantTimeInputRow components
Consolidates duplicate row components from finals, elimination, and participant
pages into two shared components with stable testids and a shared TaTimeInputProps
alias, addressing review findings #1917 #1919 #1921 #1923 #1929 #1934.

- Extract TaTimeEntryRow (shared by finals + elimination phases); optional
  livesLabel prop renders heart icons only for finals, omitted in elimination
- Extract TaParticipantTimeInputRow from participant/page.tsx
- Wrap both in memo(function Name) for stable display name without redundant
  .displayName assignment
- Rename data-testids from ta-finals-round-* to ta-time-entry-* so they
  remain accurate when used in elimination phase context; update e2e/tc-ta.js
- Use named ReactNode import instead of namespace-style import type React
