+++ smkc-score-app/src/app/tournaments/[id]/ta/participant/page.tsx
@@ -25,14 +25,13 @@
  * - Namespaces: participant, ta, common
  */
 
-import { memo, useState, useEffect, useCallback, use, useMemo } from 'react';
+import { useState, useEffect, useCallback, use, useMemo } from 'react';
 import { useTranslations } from 'next-intl';
 import { useSession } from 'next-auth/react';
 import { usePolling } from '@/lib/hooks/usePolling';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
+import { TaParticipantTimeInputRow } from '@/components/tournament/ta-participant-time-input-row';
 import { Alert, AlertDescription } from '@/components/ui/alert';
 import { AlertTriangle, Trophy, Users, Timer, LogIn, Dice5, Lock } from 'lucide-react';
 import Link from 'next/link';
@@ -78,45 +77,6 @@ interface TAApiData {
   taPlayerSelfEdit?: boolean;
 }
 
-type TaParticipantTimeInputRowProps = {
-  courseAbbr: string;
-  value: string;
-  placeholder: string;
-  disabled: boolean;
-  inputClassName: string;
-  timeInputProps: TaTimeInputProps;
-  onChange: (course: string, value: string) => void;
-  onBlur: (course: string) => void;
-};
-
-export const TaParticipantTimeInputRow = memo(function TaParticipantTimeInputRow({
-  courseAbbr,
-  value,
-  placeholder,
-  disabled,
-  inputClassName,
-  timeInputProps,
-  onChange,
-  onBlur,
-}: TaParticipantTimeInputRowProps) {
-  return (
-    <div className="flex items-center gap-2">
-      <Label className="w-12 text-xs font-mono">{courseAbbr}</Label>
-      <Input
-        type="text"
-        {...timeInputProps}
-        placeholder={placeholder}
-        value={value}
-        onChange={(e) => onChange(courseAbbr, e.target.value)}
-        onBlur={() => onBlur(courseAbbr)}
-        disabled={disabled}
-        className={inputClassName}
-      />
-    </div>
-  );
-});
-
-TaParticipantTimeInputRow.displayName = 'TaParticipantTimeInputRow';
 
 /**
  * Convert display time string to milliseconds for preview calculation.
